Indian_AI_Spectrum/Job_Hunt/backend/requirements.txt
# Job Hunt Portal - Backend Dependencies

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
starlette>=0.27.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP & Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0
lxml>=4.9.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# Authentication (for future)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6

# Email
emails>=0.6
aiosmtplib>=3.0.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
